# Make LinearSearchJumpTable use span.Equals which is much faster than string.Compare