--- a/src/Http/Routing/src/Matching/LinearSearchJumpTable.cs
+++ b/src/Http/Routing/src/Matching/LinearSearchJumpTable.cs
@@ -30,17 +30,11 @@ internal sealed class LinearSearchJumpTable : JumpTable
         }
 
         var entries = _entries;
+        var pathSpan = path.AsSpan(segment.Start..(segment.Start + segment.Length));
         for (var i = 0; i < entries.Length; i++)
         {
             var text = entries[i].text;
-            if (segment.Length == text.Length &&
-                string.Compare(
-                    path,
-                    segment.Start,
-                    text,
-                    0,
-                    segment.Length,
-                    StringComparison.OrdinalIgnoreCase) == 0)
+            if (pathSpan.Equals(text, StringComparison.OrdinalIgnoreCase))
             {
                 return entries[i].destination;
             }