--- a/src/Http/Routing/src/Matching/SingleEntryJumpTable.cs
+++ b/src/Http/Routing/src/Matching/SingleEntryJumpTable.cs
@@ -24,19 +24,14 @@ internal sealed class SingleEntryJumpTable : JumpTable
 
     public override int GetDestination(string path, PathSegment segment)
     {
-        if (segment.Length == 0)
+        var length = segment.Length;
+        if (length == 0)
         {
             return _exitDestination;
         }
 
-        if (segment.Length == _text.Length &&
-            string.Compare(
-                path,
-                segment.Start,
-                _text,
-                0,
-                segment.Length,
-                StringComparison.OrdinalIgnoreCase) == 0)
+        var pathSpan = path.AsSpan(segment.Start, length);
+        if (pathSpan.Equals(_text, StringComparison.OrdinalIgnoreCase))
         {
             return _destination;
         }