@andrewjsaid 2024-05-15
One thing I'd like to research when I have time to run the perf tests is whether `SingleEntryAsciiJumpTable` is now actually slower than `SingleEntryJumpTable` and an anti-optimization because very quick rough-and-dirty tests indicate `span.Equals(..., OrdinalIgnoreCase)` is actually faster than `Ascii.EqualsIgnoreCase(span, span)`.

i.e. deleting `SingleEntryAsciiJumpTable` would probably simplify the code and be faster.